+++ src/main/java/com/beeswork/balanceaccountservice/entity/account/AccountQuestionId.java
@@ -19,6 +19,6 @@ import java.util.UUID;
 public class AccountQuestionId implements Serializable {
 
     private UUID accountId;
-    private Long questionId;
+    private Integer questionId;
 
 }
